Support automated releases using GitHub Actions (#183)
The release binaries for GitHub are not built as part of the release process. This means they need to be compiled on a developer machine, which isn't ideal. Using GitHub Actions we can automate this process with a single workflow (+ goreleaser).

This PRs adds a workflow to GitHub Actions that will build the binaries when a release is created. This will include a windows and linux release (for amd64), with a checksums file. GoReleaser handles the compilation and deployment to github.