Remove Coveralls from the repository (#186)
While implementing the GitHub Actions work, I noticed that coveralls seems to fail often. I suspect this is a requirement (or setting) that enforces code coverage of changes. However it is throwing failures even in the case of document / build harness changes.

This PR removes coveralls and all mentions of it from the repository.