CI: Implement release creation (#260)
### Why

We want a way to build and publish new versions for non d2lers. On a
repository dispatch trigger, this will create a new release, compile BMX
for all supported operating systems and upload as assets to the release

### How

The repository dispatch trigger will be invoked from a private
brightspace repo. We use the github cli to create a release with the
version specified in the payload and then on each OS build, add a step
to upload the built project



https://github.com/Brightspace/bmx/releases/tag/untagged-30a277809d529269ae8b
example of what a release would look like. I wish the names could all be
what they are meant to be. Either just bmx or D2L.Bmx. Think they all
need unique names and couldn't find a way around that. Unless we zipped
each OS up and changed the name on the zip file?

I did not invoke this trigger to have it create the release linked. It
was done by just setting the release tag to v3.0.0 in code. So will need
to test that and confirm it still works the intended way

Related Trello card:
https://trello.com/c/k8YEuX06/1226-implement-release-strategy

---------

Co-authored-by: Chenfeng Bao <chenfeng.bao@d2l.com>