Set partial trimming (#265)
### Why

The published build of bmx right now will crash with the error `No
parameterless constructor defined for type
'Amazon.Util.Internal.PlatformServices.NetworkReachability'`


https://learn.microsoft.com/en-us/dotnet/core/deploying/trimming/trim-self-contained

When building bmx with aot, we reduce the deployment size by getting rid
of unused parts of libraries. Theres a [problem
](https://github.com/aws/aws-sdk-net/issues/2531) though with the aws
sdk right now where it gets rid of some parts we actually need. I think
it's to do with
[this](https://learn.microsoft.com/en-us/dotnet/core/deploying/trimming/incompatibilities#dynamic-assembly-loading-and-execution)

### How

The github issue page mentioned a temporary solution to set the trim
mode to partial for now. The build size would be 17.5MB compared to 13
so it seems like not a big issue?