print: support json / aws credentials output (#267)
### Why
In support of this request from a forked version of bmx
https://github.com/Brightspace/bmx/issues/204
This allows us to add entries to our ~.aws/credentials files where we
can call bmx to retrieve credentials for us.

### How
I'm adding an extra option value for --output flag as json but also want
a new flag --headless .

My thought was maybe someone could want json output outside of the
automated process? With the headless flag on it would just throw an
error anytime a needed value wasn't provided. Although it can't actually
be a headless process because if the okta session is expired we can't
get credentials. Right now with the headless flag and we don't have
credentials a non-zero exit code is returned and aws knows bmx failed

Do we want to have it split like this or just anytime the output is json
(or name it aws or something else) we don't bother looking for inputs.
Right now if you try to use an aws command without a session and without
throwing an error you don't even get prompted to enter the okta password
it just stalls for awhile
This is what an entry in the credentials file looks like
> [bmxtest]
credential_process = bmx print --headless --output json --account
Dev-Slims --role Dev-Slims-Readonlythey should add the following to
their credentials file

Related trello card:
https://trello.com/c/nOA2khxF/1230-support-awscredentials