make classes non-static where appropriate (#294)
Code that directly interacts with external systems should generally be
put behind interfaces and not be in static classes/methods. This is
necessary to have any sort of unit tests. The .NET bmx doesn't have unit
tests right now, but we should still follow this general principle as a
best practice.

Main changes:

* Consolidate all user console prompts into `ConsolePrompter` and put it
behind an interface.
* Write all console prompts to stderr instead of stdout to make `bmx
print` usable.
* Put `OktaSessionStorage` behind an interface too
* Rename `Authenticator` to `OktaAuthenticator` and move it to the Okta
namespace

Note: I'm not putting all classes behind interfaces (e.g.
`OktaAuthenticator`). Even when we want to unit test the code, code that
doesn't directly interact with external systems doesn't necessarily need
to be mocked. We can introduce interfaces when/if such a need arises.