remove allow_project_configs config option (#323)
### Why

Slack discussion
[here](https://d2l.slack.com/archives/G4P099831/p1685471844688699)

tldr: This feature came in the Go version without previously being in V1
and guessing we left it as an experimental flag. Partly because it could
crash on Linux systems working its way up the directories. This isn't a
problem now in .NET version.

This feature has matured enough to become a default setting. So now we
will simply always look for a .bmx file in the CWD.