write: use ini parser to format credentials file (#333)
### Why

The `SharedCredentialsFile` has a good implementation for working with
ini files but the output is that of `x=y`. The GO version of BMX had the
output such that every `=` sign for each line was aligned.

### How

Use a third party ini parser library that will let us update / create
profiles. The libraries I looked through don't have this same feature
for dotnet.

I considered writing something custom to give us the same kind of output
here. While I got that working, I was worried of edge cases where the
time spent wouldn't be worth it. Last thing I want is to break someone's
credentials file.

So the new output will be in the format `x = y` instead of `x=y`