check for updates on run (#397)
### Why

Helps users know we have a new version out. This is a middle step in
terms of having an update feature. It'd be good to keep all our users on
the latest version.

We will ping github once every 24 hours and save the version tag it
returned. So currently if there is a new version the user will see it
every time they run BMX. I could change that to once every 24 hours
though but this also maybe annoys them enough to want to upgrade lol

If the org is d2l we prompt them to visit out site with the install
script. Users outside of d2l will get prompted to visit the github repo

![image](https://github.com/Brightspace/bmx/assets/90227099/63570694-4905-4fb7-8f5b-47dcb856f178)

![image](https://github.com/Brightspace/bmx/assets/90227099/f47e8966-3381-4256-a17a-e4f07d1f77bc)

related trello card: https://trello.com/b/r6wdvBry/vulcan-project-board
slack discussion:
https://d2l.slack.com/archives/G4NV94H2A/p1693323866881979

---------

Co-authored-by: github-actions[bot] <41898282+github-actions[bot]@users.noreply.github.com>
Co-authored-by: DotNet Format Bot <dotnet-format-bot@github.d2l>
Co-authored-by: Adipa Wijayathilaka <adipamihi@gmail.com>