write: update config file section with same name if no cred process flag (#453)
### Why

Default credentials file takes priority over the default config which is
where we store profiles using `credential_process`. If someone had an
existing profile in config file and then runs `bmx write`, creds will be
stored into the default credentials file making the existing one
useless. We should remove the existing key in the config file so there
is no confusion.

`credential_process` is the better way for majority of people but there
are some scenarios where it's not supported. So we are leaving some
information to the user that it is no longer being used. Just in case
they unintentionally ran write without the flag. Not going with a prompt
to confirm because that could break existing scripts

### Ticket

VUL-385

---------

Co-authored-by: Chenfeng Bao <chenfeng.bao@d2l.com>