tweak release workflow (#490)
* We never published a release automatically, and I don't think we ever
should.
We need to manually edit and review the release, and test the
to-be-released binaries (as opposed to testing binaries built from
another CI run or even locally).
  So I've changed the workflow to keep (non-pre-)releases as draft.

* I'd like prereleases to all look like prereleases.
The current process doesn't enforce that you must include words like
"preview", "beta" when you publish a prerelease. This can be confusing.
I've made the workflow auto append "-preview-{datetime}" to the
version/tag name for prereleases.