Add AWS_PROFILE env var to readme (#507)
The info in my change is obvious to some folks, but I lost a good chunk
of time to this and it seems like a straightforward addition.

Essentially when I read the readme I only saw a way to use profiles when
invoking AWS through the CLI (`--profile` option), whereas my usage was
inside Node scripts. I just wanted env var(s) I could set so that my
Node scripts would work. I ended up trying to go the `bmx print` route
but had lots of grief trying to do that.

I also thought the profile was a BMX concept and not AWS, which is my
fault but hopefully this makes it clearer!